Integrated Google OAuth authentication with REST framework

+++ pkg/app/views.py
@@ -1,12 +1,17 @@
-from django.shortcuts import render
+from django.http import HttpResponseRedirect
+from django.contrib.auth import logout as auth_logout
+from django.contrib.auth.decorators import login_required
+from django.template import RequestContext
+from django.shortcuts import render, render_to_response, redirect
+from django.contrib.messages.api import get_messages
 
 from app.models import DataSet, Resource, Tag, DataSetExtra, Group
-from app.models import DataSetResource
+from app.models import DataSetResource, DataSetRating
 from app.serializers import DataSetSerializer, ResourceSerializer
 from app.serializers import TagSerializer, DataSetExtraSerializer
 from app.serializers import GroupSerializer
 
-from rest_framework import generics, status
+from rest_framework import generics, status, permissions
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework.reverse import reverse
@@ -29,30 +34,37 @@ def api_root(request, format=None):
 class DataSetList(generics.ListCreateAPIView):
     queryset = DataSet.objects.all()
     serializer_class = DataSetSerializer
+    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
 
 class DataSetDetail(generics.RetrieveUpdateDestroyAPIView):
     queryset = DataSet.objects.all()
     serializer_class = DataSetSerializer
+    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
 
 class ResourceList(generics.ListCreateAPIView):
     queryset = Resource.objects.all()
     serializer_class = ResourceSerializer
+    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
 
 class ResourceDetail(generics.RetrieveUpdateDestroyAPIView):
     queryset = Resource.objects.all()
     serializer_class = ResourceSerializer
+    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
 
 class GroupList(generics.ListCreateAPIView):
     queryset = Group.objects.all()
     serializer_class = GroupSerializer
+    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
 
 class GroupDetail(generics.RetrieveUpdateDestroyAPIView):
     queryset = Group.objects.all()
     serializer_class = GroupSerializer
+    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
 
 class TagList(generics.ListCreateAPIView):
     queryset = Tag.objects.all()
     serializer_class = TagSerializer
+    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
 
 
 
@@ -93,6 +105,9 @@ class DataSetResourceDetail(APIView):
             raise Http404
 
     def post(self, request, pk, pk_res, format=None):
+        if not request.user.is_active:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         dataset = self.get_object(pk)
         resource = self.get_resource(pk_res)
         resources = DataSetResource.objects.filter(dataset=dataset)
@@ -104,6 +119,9 @@ class DataSetResourceDetail(APIView):
         return Response(status=status.HTTP_201_CREATED)
 
     def delete(self, request, pk, pk_res, format=None):
+        if not request.user.is_active:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         dataset = self.get_object(pk)
         resource = self.get_resource(pk_res)
         #dsr = self.get_dsr(pk, pk_res)
@@ -137,6 +155,9 @@ class DataSetExtraList(APIView):
         return Response(serializer.data)
 
     def post(self, request, pk, format=None):
+        if not request.user.is_active:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         dataset = self.get_object(pk)
 
         serializer = DataSetExtraSerializer(data=request.data)
@@ -167,6 +188,9 @@ class DataSetExtraDetail(APIView):
             raise Http404
 
     def put(self, request, pk, pk_extra, format=None):
+        if not request.user.is_active:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         serializer = DataSetExtraSerializer(
             self.get_extra(pk, pk_extra), data=request.data)
         if serializer.is_valid():
@@ -175,5 +199,56 @@ class DataSetExtraDetail(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk, pk_extra, format=None):
+        if not request.user.is_active:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         self.get_extra(pk, pk_extra).delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+
+##
+# Data Set Rater. Handles rating a dataset
+##
+class DataSetRate(APIView):
+    def get_object(self, pk):
+        try:
+            return DataSet.objects.get(pk=pk)
+        except DataSet.DoesNotExist:
+            raise Http404
+
+    def get(self, request, pk, rating):
+        if not request.user.is_active:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
+        dataset = self.get_object(pk)
+        try:
+            dr = dataset.datasetrating_set.get(user=request.user)
+            dr.rating = rating
+            dr.save()
+        except DataSetRating.DoesNotExist:
+            dr = dataset.datasetrating_set.create(user=request.user, rating=rating)
+            dr.save()
+        return Response(status=status.HTTP_201_CREATED)
+
+
+###
+# OAuth Views
+###
+@login_required
+def logged_in(request):
+    """Login complete view, displays user data"""
+    ctx = {
+        'last_login': request.session.get('social_auth_last_login_backend'),
+    }
+    return render_to_response('logged-in.html', ctx, RequestContext(request))
+
+def login_error(request):
+    messages = get_messages(request)
+    return render_to_response('login-error.html', {'version': version,
+        'messages': messages},
+        RequestContext(request))
+
+def logout(request):
+    auth_logout(request)
+    return HttpResponseRedirect('/')
